Extract EventPopup component and refactor EventMap
Moved marker/popup rendering into a dedicated EventPopup component.
